final cleanup and code stacking

+++ Matrix Classes/src/packages/abstractMatrixPackage/MatrixAbstract.hpp
@@ -43,39 +43,39 @@ namespace matrixAbstractPackage {
    */
   template<typename M>
   class MatrixAbstract {
-  public:
-    MatrixAbstract();
-    MatrixAbstract(unsigned short int&, unsigned short int&); /** Sygnatura konstr. dla macierzy prostokątnych */
-    MatrixAbstract(unsigned short int&); /** Sygnatura konstr. dla macierzy kwadratowych */
-    MatrixAbstract(const MatrixAbstract<M>&); /** Sygnatura konstr. kopiującego */
+    public:
+      MatrixAbstract();
+      MatrixAbstract(unsigned short int&, unsigned short int&); /** Sygnatura konstr. dla macierzy prostokątnych */
+      MatrixAbstract(unsigned short int&); /** Sygnatura konstr. dla macierzy kwadratowych */
+      MatrixAbstract(const MatrixAbstract<M>&); /** Sygnatura konstr. kopiującego */
 
-    void printMtrx(HANDLE&, const bool, const bool, const bool) const; /** Drukowanie macierzy */
-    double scalarValuePush(HANDLE&); /** Wprowadzanie wartości skalara */
+      void printMtrx(HANDLE&, const bool, const bool, const bool) const; /** Drukowanie macierzy */
+      double scalarValuePush(HANDLE&); /** Wprowadzanie wartości skalara */
 
-    virtual void mtrxTypeAndSizeInfo() = 0; /** Wypisywanie wiadomosci */
-    virtual void insertMtrx(HANDLE&) = 0; /** Wprowadzanie macierzy */
+      virtual void mtrxTypeAndSizeInfo() = 0; /** Wypisywanie wiadomosci */
+      virtual void insertMtrx(HANDLE&) = 0; /** Wprowadzanie macierzy */
 
-    /** Metody statyczne (dodatek interfejsu użytkownika) */
-    static void genInfoBlock(std::string, std::vector<std::string>);
-    static void sequentialMess(unsigned int, std::string);
-    static void errorMess(std::string, HANDLE&);
+      /** Metody statyczne (dodatek interfejsu użytkownika) */
+      static void genInfoBlock(std::string, std::vector<std::string>);
+      static void sequentialMess(unsigned short int, std::string);
+      static void errorMess(std::string, HANDLE&);
 
-    virtual ~MatrixAbstract(); /** Wirtualny destruktor wywołujący destruktory z klas pochodnych */
+      virtual ~MatrixAbstract(); /** Wirtualny destruktor wywołujący destruktory z klas pochodnych */
 
-  private: /** Metody prywatne; dostępne tylko na użytek metod klasy abstrakcyjnej */
-    void allocateMemory(); /** Alokacja pamięci */
-    unsigned int findMaxLength(unsigned int&) const;
-    unsigned int lengthOfElm(M&) const;
+    private: /** Metody prywatne; dostępne tylko na użytek metod klasy abstrakcyjnej */
+      void allocateMemory(); /** Alokacja pamięci */
+      unsigned short int findMaxLength(unsigned short int&) const;
+      unsigned short int lengthOfElm(M&) const;
 
-  public: /** Gettery */
-    unsigned short int get_Cols() const;
-    unsigned short int get_Rows() const;
-    M** get_Mtrx() const;
+    public: /** Gettery */
+      unsigned short int get_Cols() const;
+      unsigned short int get_Rows() const;
+      M** get_Mtrx() const;
 
-  protected:
-    unsigned short int mtrxWidth{0}, mtrxHeight{0}; /** Wymiary macierzy */
-    double scalarVal{0}; /** Wartość skalara */
-    M** mtrx{nullptr}; /** Tablica wskaźników typu M reprezentująca macierz */
+    protected:
+      unsigned short int mtrxWidth{0}, mtrxHeight{0}; /** Wymiary macierzy */
+      double scalarVal{0}; /** Wartość skalara */
+      M** mtrx{nullptr}; /** Tablica wskaźników typu M reprezentująca macierz */
   };
 
   #include "src/constructors.tpp" /** Deklaracje konstruktorów */
